Extract startupChecks, fold catchPrintOutput into rubinTv, delete utils.py
Final commit of the utils.py split:

- Move setupSentry and checkRubinTvExternalPackages out of utils.py
  into a new startupChecks module. These two functions are called
  near the top of every script wrapper to make sure the process is
  in a runnable state before it does anything; grouping them in
  their own small module makes the intent obvious. The
  EFD_CLIENT_MISSING_MSG and GOOGLE_CLOUD_MISSING_MSG strings move
  with them.
- Move catchPrintOutput into rubinTv.py as a private
  `_catchPrintOutput` helper. It is used in exactly one place — the
  NightReportChannel `printShutterTimes`/`printObsGaps` capture —
  and is small enough that hoisting it to its own module would be
  pure overhead. Renaming it to a private leading-underscore name
  makes its single-use scope explicit.
- Delete utils.py outright. With the predicates / formatters /
  parsers / shardIo / locationConfig / butlerQueries / timing /
  startupChecks splits all landed, plus catchPrintOutput now in
  its sole consumer, there is nothing left in utils.py. As per the
  "this is an application, not a library" convention, no
  compatibility shim is left behind: the file is gone.

All call sites for setupSentry / checkRubinTvExternalPackages have
been moved to import from startupChecks instead of utils. After
this commit, no file in the package, scripts, or tests directory
references `from .utils import …` or
`from lsst.rubintv.production.utils import …` — verified by grep.

Total picture of the utils.py split (across all eight commits):
the original 1,889-line `utils.py` has been reorganised into eight
focused modules (`predicates`, `formatters`, `parsers`, `shardIo`,
`locationConfig`, `butlerQueries`, `timing` extension,
`startupChecks`), each one importable in isolation. Every internal
caller across the package, the scripts directory, and the test
suite has been migrated to import directly from the new modules.

Verified:
- mypy and flake8 clean across the package, scripts, and tests.
- The unit test suite (192 tests, +157 since the start of the
  branch) passes. The two pre-existing environment-dependent
  failures (test_metadataService and test_pipelines, which need a
  fully-configured RAPID_ANALYSIS_LOCATION on the host) are
  unaffected by the refactor.
- Every dependent in-package module imports cleanly.
- A representative selection of scripts parses and imports
  cleanly.

Written by Claude Code 4.6 Opus (high-effort).
