+++ scripts/LSSTCam/runButlerCanary.py
@@ -24,7 +24,7 @@ import time
 
 from lsst.daf.butler import Butler, DataCoordinate
 from lsst.rubintv.production.locationConfig import getAutomaticLocationConfig
-from lsst.rubintv.production.utils import setupSentry
+from lsst.rubintv.production.startupChecks import setupSentry
 from lsst.summit.utils.utils import setupLogging
 
 setupSentry()
