+++ scripts/LSSTCam/runClusterManager.py
@@ -23,7 +23,7 @@ from lsst.daf.butler import Butler
 from lsst.rubintv.production.clusterManagement import ClusterManager
 from lsst.rubintv.production.locationConfig import getAutomaticLocationConfig
 from lsst.rubintv.production.predicates import getDoRaise
-from lsst.rubintv.production.utils import setupSentry
+from lsst.rubintv.production.startupChecks import setupSentry
 from lsst.summit.utils.utils import setupLogging
 
 setupSentry()
