+++ scripts/LSSTCam/runOneOffPostIsr.py
@@ -24,7 +24,7 @@ from lsst.rubintv.production.locationConfig import getAutomaticLocationConfig
 from lsst.rubintv.production.oneOffProcessing import OneOffProcessor
 from lsst.rubintv.production.podDefinition import PodDetails, PodFlavor
 from lsst.rubintv.production.predicates import getDoRaise
-from lsst.rubintv.production.utils import setupSentry
+from lsst.rubintv.production.startupChecks import setupSentry
 from lsst.summit.utils.utils import setupLogging
 
 setupSentry()
