+++ scripts/LSSTCam/runRadialPlotting.py
@@ -22,7 +22,7 @@
 from lsst.daf.butler import Butler
 from lsst.rubintv.production.aos import RadialPlotter
 from lsst.rubintv.production.locationConfig import getAutomaticLocationConfig
-from lsst.rubintv.production.utils import setupSentry
+from lsst.rubintv.production.startupChecks import setupSentry
 from lsst.summit.utils.utils import setupLogging
 
 setupSentry()
