+++ scripts/LSSTCam/runSfmRunner.py
@@ -29,7 +29,7 @@ from lsst.rubintv.production.locationConfig import getAutomaticLocationConfig
 from lsst.rubintv.production.pipelineRunning import SingleCorePipelineRunner
 from lsst.rubintv.production.podDefinition import PodDetails, PodFlavor
 from lsst.rubintv.production.predicates import getDoRaise
-from lsst.rubintv.production.utils import setupSentry
+from lsst.rubintv.production.startupChecks import setupSentry
 from lsst.summit.utils.utils import setupLogging
 
 setupSentry()
