+++ scripts/summit/auxTel/runCalibrateCcdRunner.py
@@ -23,7 +23,7 @@ import sys
 
 from lsst.rubintv.production import CalibrateCcdRunner
 from lsst.rubintv.production.locationConfig import LocationConfig
-from lsst.rubintv.production.utils import checkRubinTvExternalPackages
+from lsst.rubintv.production.startupChecks import checkRubinTvExternalPackages
 from lsst.summit.utils.utils import setupLogging
 
 setupLogging()
