+++ scripts/summit/misc/runStarTrackerMetadata.py
@@ -22,8 +22,8 @@
 import sys
 
 from lsst.rubintv.production.locationConfig import LocationConfig
+from lsst.rubintv.production.startupChecks import checkRubinTvExternalPackages
 from lsst.rubintv.production.timedServices import TimedMetadataServer
-from lsst.rubintv.production.utils import checkRubinTvExternalPackages
 from lsst.summit.utils.utils import setupLogging
 
 setupLogging()
